api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
stripe==11.3.0
email-validator==2.2.0
httpx==0.27.2
cantera==3.2.0
numpy==2.1.3
scipy==1.14.1
cryptography==43.0.3
